<feat>: HomeFragment, scheduleDeleteDialog의 api 기초공사

+++ project/Myo_jib_sa/app/src/main/java/com/example/myo_jib_sa/schedule/api/scheduleOfDay/ScheduleOfDayResponse.kt
@@ -8,7 +8,7 @@ data class ScheduleOfDayResponse(
     val result: List<ScheduleOfDayResult>
 )
 data class ScheduleOfDayResult(
-    val scheduleId: Int,
+    val scheduleId: Long,
     val scheduleTitle: String,
     val scheduleStart: String,
     val scheduleEnd: String,
